Update Macro_I_Creating_selections_cell_counting.ijm
This macro iterates through all files in a folder (seriated 2D slices of a brain with labelled cells) and asks the user to create selections (ROIs) around the cell-containing regions.

It requires a folder containing all the images to be analyzed. Images have to be 1 channel .tiff format. Minimal modifications are required for multichannel (composite) images

It will create a folder containing the specified ROI for each slice that will be used in Macro_II

//Questions @ https://github.com/JRamirez-F/   or open a topic in https://forum.image.sc/ and mention @J_Ramirez